+++ library/src/main/java/org/dbtools/android/commons/ui/compose/dialog/MessageDialog.kt
@@ -21,9 +21,9 @@ fun MessageDialog(
     text: String? = null,
     icon: @Composable (() -> Unit)? = null,
     confirmButtonText: @Composable () -> String? = { stringResource(android.R.string.ok) },
-    onConfirmButtonClicked: (() -> Unit)? = null,
+    onConfirmButtonClick: (() -> Unit)? = null,
     dismissButtonText: @Composable () -> String? = { stringResource(android.R.string.cancel) },
-    onDismissButtonClicked: (() -> Unit)? = null,
+    onDismissButtonClick: (() -> Unit)? = null,
     tonalElevation: Dp = AlertDialogDefaults.TonalElevation
 ) {
     require(title != null || text != null) { "Title or Text is required (if visible)" }
@@ -43,10 +43,10 @@ fun MessageDialog(
         icon = icon,
         confirmButton = {
             val confirmButtonTextString = confirmButtonText()
-            if (onConfirmButtonClicked != null && confirmButtonTextString != null) {
+            if (onConfirmButtonClick != null && confirmButtonTextString != null) {
                 TextButton(
                     onClick = {
-                        onConfirmButtonClicked()
+                        onConfirmButtonClick()
                     }
                 ) {
                     Text(confirmButtonTextString)
@@ -55,10 +55,10 @@ fun MessageDialog(
         },
         dismissButton = {
             val dismissButtonTextString = dismissButtonText()
-            if (onDismissButtonClicked != null && dismissButtonTextString != null) {
+            if (onDismissButtonClick != null && dismissButtonTextString != null) {
                 TextButton(
                     onClick = {
-                        onDismissButtonClicked()
+                        onDismissButtonClick()
                     }
                 ) {
                     Text(dismissButtonTextString)
@@ -79,9 +79,9 @@ fun MessageDialog(
         text = dialogUiState.text?.invoke(),
         icon = dialogUiState.icon,
         confirmButtonText = dialogUiState.confirmButtonText,
-        onConfirmButtonClicked = if (dialogUiState.onConfirm != null) { { dialogUiState.onConfirm.invoke(Unit) } } else null,
+        onConfirmButtonClick = if (dialogUiState.onConfirm != null) { { dialogUiState.onConfirm.invoke(Unit) } } else null,
         dismissButtonText = dialogUiState.dismissButtonText,
-        onDismissButtonClicked = dialogUiState.onDismiss,
+        onDismissButtonClick = dialogUiState.onDismiss,
     )
 }
 
@@ -104,8 +104,8 @@ private fun PreviewMessageDialog() {
             title = "Title",
             text = "This is the content that goes in the text",
             onDismissRequest = { },
-            onConfirmButtonClicked = { },
-            onDismissButtonClicked = { }
+            onConfirmButtonClick = { },
+            onDismissButtonClick = { }
         )
     }
 }
@@ -119,8 +119,8 @@ private fun PreviewMessageDialogWithIcon() {
             text = "This is the content that goes in the text",
             icon = { Icon(imageVector = Icons.Default.CheckCircle, contentDescription = null) },
             onDismissRequest = { },
-            onConfirmButtonClicked = { },
-            onDismissButtonClicked = { }
+            onConfirmButtonClick = { },
+            onDismissButtonClick = { }
         )
     }
 }